+++ src/vc4.h
@@ -78,6 +78,6 @@ extern int unity_kernel;
 extern int kernel_start;
 
 int compute_nearest_neighbour_kernel(uint32_t *dlist_memory);
-int compute_scaling_kernel(uint32_t *dlist_memory, double b, double c);
+int compute_scaling_kernel(uint32_t *dlist_memory, ULONG b, ULONG c);
 
 #endif /* _VC4_H */
